Add Scene for Characters and create prefab
