+++ Assets/Prefabs/Client/UI/CharacterPickerUI.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 55d63129ef1fd7d4ca9e0acb7141d022
-folderAsset: yes
-DefaultImporter:
+guid: 3175c1cf3a336634aaf88c4c5e255ef0
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
