+++ Assets/Scenes/Client/ClientMyCharacters.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 5a2e144eb6c477743af7dec12d434c7a
-folderAsset: yes
+guid: 2ef865aebd5e9df4c8667eb784ac8612
 DefaultImporter:
   externalObjects: {}
   userData:
