+++ Assets/Scripts/Client/ClientSyncStates/ClientMyCharactersState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8216e9f3f6ed9e742ac63d422757ad02
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
