+++ Assets/Scripts/UI/Client/CharacterPickerUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 643775a18f84a7b4bb4f4587c34e38c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
