+++ Assets/Scripts/UI/Client/ClientCharacterSelectUI.cs
@@ -1,11 +1,5 @@
 ﻿using UnityEngine;
-using System.Collections;
-using System.Collections.Generic;
-using TMPro;
-using static ubv.microservices.CharacterDataService;
-using UnityEngine.SceneManagement;
 using UnityEngine.UI;
-using static ubv.client.logic.ClientCharacterSelectState;
 using UnityEngine.EventSystems;
 
 namespace ubv.ui.client
@@ -15,19 +9,11 @@ namespace ubv.ui.client
         [SerializeField] private ubv.client.logic.ClientCharacterSelectState m_initState;
         [SerializeField] private Button m_searchButton;
         [SerializeField] private AddCharacterUI m_addCharacterForm;
-        [SerializeField] private TextMeshProUGUI m_characterName;
-        [SerializeField] private Transform m_countDotListParent;
-        [SerializeField] private Image m_defaultCountDotItem;
-        [SerializeField] private Color m_normalDotColor;
-        [SerializeField] private Color m_selectedDotColor;
-        private List<Image> m_characterCountDot;
         private EventSystem system;
 
         private void Awake()
         {
             system = EventSystem.current;
-            m_characterCountDot = new List<Image>();
-            m_characterCountDot.Add(m_defaultCountDotItem);
         }
 
         protected override void Update()
@@ -36,8 +22,6 @@ namespace ubv.ui.client
 
             if (Time.frameCount % 69 == 0)
             {
-                m_characterName.text = m_initState.GetActiveCharacter()?.Name;
-                SetSelectionDotColor(m_initState.GetCharacterCount());
                 if (system.currentSelectedGameObject == null)
                 {
                     m_searchButton.Select();
@@ -45,66 +29,10 @@ namespace ubv.ui.client
             }
         }
 
-        public void NextCharacter()
-        {
-            string characterName = m_initState.NextCharacter();
-            if (characterName != null)
-            {
-                m_characterName.text = characterName;
-                SetSelectionDotColor(m_initState.GetCharacterCount());
-            }  
-        }
-
-        public void PreviousCharacter()
-        {
-            string characterName = m_initState.PreviousCharacter();
-            if (characterName != null)
-            {
-                m_characterName.text = characterName;
-                SetSelectionDotColor(m_initState.GetCharacterCount());
-            }
-        }
-
-        public void DeleteActiveCharacter()
-        {
-            m_initState.DeleteActiveCharacter();
-        }
 
         public void Back()
         {
             m_initState.Back();
-        }
-
-        private void SetSelectionDotColor(CharacterCount characterCount)
-        {
-            while (m_characterCountDot.Count != characterCount.total)
-            {
-                if(m_characterCountDot.Count > characterCount.total)
-                {
-                    int lastIndex = m_characterCountDot.Count - 1;
-                    Destroy(m_characterCountDot[lastIndex].gameObject);
-                    m_characterCountDot.RemoveAt(lastIndex);
-                }
-                else if(m_characterCountDot.Count < characterCount.total)
-                {
-                    Image playerItem = GameObject.Instantiate(m_defaultCountDotItem, m_countDotListParent);
-                    m_characterCountDot.Add(playerItem);
-                }
-                else
-                {
-                    break;
-                }
-            }
-
-            for (int i = 0; i < m_characterCountDot.Count; i++)
-            {
-                Color color = m_normalDotColor;
-                if(i == characterCount.selected)
-                {
-                    color = m_selectedDotColor;
-                }
-                m_characterCountDot[i].color = color;
-            }
         }
     }
 }
