+++ Assets/Scripts/UI/Client/ClientMyCharactersUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5b50a6aceb08f64fbb1362aa4885e14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
